Don't scope EEx layout templates - wire up the opts
The info isn't used yet, but this commit flags layout/* files as "not
scopable".  We only have 1 level of scoping in APM right now, and since
practically every template has a layout, it loses a more interesting
level of scope if we "spend" it on layouts. Instead now we can see when
Ecto or whatever is called from inside a template