Add option to control whether the app name is trimmed from controllers
Currently, the way the library is setup controller instrumentation
reports controller names sans the app name. As an example, in a standard
Phoenix app if I have a controller called `ExampleApp.SessionController`
then it will be reported under `SessionController#action_name`. This is
very convenient for apps that only have one phoenix app since it reduces
verbosity.

However, we're running Scout in an umbrella app that has multiple
Phoenix apps underneath it. So in our setup we might have something that
looks like this: `ExampleApp1.SessionController` and
`ExampleApp2.SessionController`. With the current implementation,
instrumentation of both controllers results in metrics being sent to the
same `SessionController#action_name` which is undesirable because we
would like them to be reported separately under their respective app
names.

In order to retain the existing behavior and avoid introducing a change
that would cause issues with anyone's existing metrics I've introduced a
configuration option that controls whether or not the app module name is
trimmed. By default the value is `true` indicating that the app module
name will be trimmed.